Added forward references for "anonymous" and "none" to discussion of signature block in 6.3.4 as requested by Stephen Farrell:
From: Stephen Farrell

Description: Section 6.4.1.1: this introduces the "anonymous" and "none" values for SignatureAndHashAlgorithm on page 88. That really needs to be introduced where signing is first discussed and you need to say when its allowed to be used and for what. (In this part you just say it cannot be used.)

Notes: Add text to 6.3.4 Security Block under "The signature construct contains the following values" . Add forward reference to 7.4.2.2 in last bullet of 7.4.1.1