Wesley Eddy's comment
The comment is:

"The document seems to be schizophrenic about Head of Line blocking.
The section 6.5.1.6 text on why TCP is not a desirable Overlay Link
Protocol notes Head of Line blocking as the primary reason to prefer
another protocol, yet the stop and wait algorithm in 6.6.3.1 for use
over DTLS, says that only one message can be unacknowledged at a time,
so it's unclear how this avoid Head of Line blocking issues (if at all).
It seems like you seay HoL issues are worth avoiding, and then define
operation only over transports with HoL issues (TCP and the DTLS-based
ones)."

Added a note saying that none of the protocols defined in the document
meets that condition, but that future protocols will.