Mary Barnes' comment 2
The text of the comment is:

"- [-17] Section 1.2.1, 2nd paragraph: I don't understand the example
as to why a single application requires multiple usages - i.e, why
voicemail?  Isn't the intent to say that an application might need to
use both SIP and XMPP - i.e., you wouldn't define a "usage" for an
application, would you?
[While Cullen responded to this comment with an explanation, there was
no change to clarify the text and Marc's response didn't help clarify
my concern]"

Cullen's response was:

"storing voicemail in the DHT would require a different usage than the
type of rendezvous service of XMPP or SIP"

- Text modified by reusing Cullent's response.