Polina Goltsman major 2
"2) sec. 7.4.1.1: replica number handling is unclear it is unclear how
replica numbers are incremented (e.g., per peer) and what receiving peers
should actually do with this number. Is it important to store the value or
would a boolean be sufficient so that the peer knows that it's a
replica?"

- Add text explaining that repliac number are allocated an interpreted
  by the topology plugin.
- Explained how CHORD-RELOAD interprets the replica numbers.